requirements.txt
requests
openai
python-dotenv
gradio==5.38.2